# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Fact Checker Application API
  description: |
    A comprehensive fact-checking platform API that provides endpoints for user authentication, 
    fact management, community interaction, moderation, and administrative functions.
    
    ## Features
    - User authentication and profile management
    - Fact submission and verification
    - Community voting and commenting
    - Content moderation and reporting
    - Analytics and insights
    - Administrative dashboard
    - UI framework components
    
    ## Authentication
    This API uses session-based authentication. Users must log in to access protected endpoints.
    
  version: 1.0.0
  contact:
    name: API Support
    email: api-support@factchecker.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000
    description: Development server
  - url: https://api.factchecker.com
    description: Production server
...
paths:
...
  /facts:
    get:
      tags: [Facts]
      summary: List facts
      description: Retrieve facts with optional filtering and pagination
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: search
          in: query
          schema:
            type: string
        - name: hashtag
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
            enum: [newest, oldest, most_voted]
            default: newest
      responses:
        '200':
          description: List of facts
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      facts:
                        type: array
                        items:
                          $ref: '#/components/schemas/Fact'
                      pagination:
                        $ref: '#/components/schemas/Pagination'

    post:
      tags: [Facts]
      summary: Create fact
      description: Submit a new fact
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content:
                  type: string
                  minLength: 10
                  maxLength: 2000
                  example: This is an important fact about science.
                resources:
                  type: array
                  items:
                    type: object
                    properties:
                      url:
                        type: string
                        format: uri
                      title:
                        type: string
      responses:
        '201':
          description: Fact created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      fact:
                        $ref: '#/components/schemas/Fact'
  /facts/{fact_id}:
    get:
      tags: [Facts]
      summary: Get fact
      description: Retrieve a specific fact by ID
      parameters:
        - name: fact_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Fact details
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      fact:
                        $ref: '#/components/schemas/FactDetail'

    put:
      tags: [Facts]
      summary: Update fact
      description: Update a fact (requires ownership)
      security:
        - sessionAuth: []
      parameters:
        - name: fact_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
                  minLength: 10
                  maxLength: 2000
      responses:
        '200':
          description: Fact updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'

    delete:
      tags: [Facts]
      summary: Delete fact
      description: Delete a fact (requires ownership or moderator role)
      security:
        - sessionAuth: []
      parameters:
        - name: fact_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Fact deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'

  # Voting Endpoints
  /voting/fact/{fact_id}:
    post:
      tags: [Voting]
      summary: Vote on fact
      description: Cast a vote on a fact (fact or fake)
      security:
        - sessionAuth: []
      parameters:
        - name: fact_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [vote_type]
              properties:
                vote_type:
                  type: string
                  enum: [fact, fake]
      responses:
        '200':
          description: Vote recorded successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      vote_counts:
                        $ref: '#/components/schemas/FactVoteCounts'
  /voting/comment/{comment_id}:
    post:
      tags: [Voting]
      summary: Vote on comment
      description: Cast a vote on a comment (upvote or downvote)
      security:
        - sessionAuth: []
      parameters:
        - name: comment_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [vote_type]
              properties:
                vote_type:
                  type: string
                  enum: [upvote, downvote]
      responses:
        '200':
          description: Vote recorded successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      vote_counts:
                        $ref: '#/components/schemas/CommentVoteCounts'

  # Comment Endpoints
...
  /comments/fact/{fact_id}:
    get:
      tags: [Comments]
      summary: Get fact comments
      description: Retrieve comments for a specific fact
      parameters:
        - name: fact_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
        - name: sort
          in: query
          schema:
            type: string
            enum: [newest, oldest, most_voted]
            default: newest
      responses:
        '200':
          description: List of comments
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      comments:
                        type: array
                        items:
                          $ref: '#/components/schemas/CommentWithReplies'
                      pagination:
                        $ref: '#/components/schemas/Pagination'

  # Admin Endpoints
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: session
  schemas:
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Operation completed successfully
...
    Fact:
      type: object
      properties:
        id:
          type: string
          format: uuid
        content:
          type: string
        user_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        vote_counts:
          $ref: '#/components/schemas/FactVoteCounts'
        hashtags:
          type: array
          items:
            type: string
        comment_count:
          type: integer
    FactDetail:
      allOf:
        - $ref: '#/components/schemas/Fact'
        - type: object
          properties:
            resources:
              type: array
              items:
                $ref: '#/components/schemas/Resource'
            comments:
              type: array
              items:
                $ref: '#/components/schemas/Comment'
    Comment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        content:
          type: string
        user_id:
          type: string
          format: uuid
        fact_id:
          type: string
          format: uuid
        parent_id:
          type: string
          format: uuid
          nullable: true
        created_at:
          type: string
          format: date-time
        vote_counts:
          $ref: '#/components/schemas/CommentVoteCounts'
    CommentWithReplies:
      allOf:
        - $ref: '#/components/schemas/Comment'
        - type: object
          properties:
            replies:
              type: array
              items:
                $ref: '#/components/schemas/Comment'
    Resource:
      type: object
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
          format: uri
        title:
          type: string
        resource_type:
          type: string
          enum: [url, image]
    FactVoteCounts:
      type: object
      properties:
        fact:
          type: integer
        fake:
          type: integer
    CommentVoteCounts:
      type: object
      properties:
        upvotes:
          type: integer
        downvotes:
          type: integer
    Pagination:
      type: object
      properties:
        page:
          type: integer
        per_page:
          type: integer
        total:
          type: integer
        pages:
          type: integer
...
security:
  - sessionAuth: []